tek1/requirements.txt
dataclasses==0.6    
Django==2.0.2  
django-crispy-forms==1.7.2
django-queryset-csv==1.0.2
djangorestframework==3.10.1 
djongo==1.2.33 
pip==19.1.1 
pymongo==3.8.0  
pytz==2018.9 
setuptools==40.6.2 
sqlparse==0.2.4  
unicodecsv==0.14.1
wheel==0.32.3 
psycopg2>=2.7,<3.0
